requirements.txt
#sudo apt-get install postgresql postgresql-contrib libpq-dev python-dev
django==1.11

pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.5


django-imagekit
django-mail_templated
xlsxwriter==0.9.3
djangorestframework==3.3.3

#====================
# DEVELOPMENT
#====================

#django-debug-panel==0.8.3
#django-debug-toolbar==1.6

#Note
#django-appconf==1.0.3


#Run this in Python 2.7
#mkvirtualenv -p /usr/bin/python2.7 <ENV_NAME>
